Move BTreeKeyData class definition out to separate header file.
