+++ searchlib/src/vespa/searchlib/btree/CMakeLists.txt
@@ -1,6 +1,7 @@
 # Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 vespa_add_library(searchlib_btree OBJECT
     SOURCES
+    btree_key_data.cpp
     btreeaggregator.cpp
     btreebuilder.cpp
     btreeinserter.cpp
