+++ searchlib/src/vespa/searchlib/btree/btreenode.cpp
@@ -5,8 +5,6 @@
 
 namespace search::btree {
 
-BTreeNoLeafData BTreeNoLeafData::_instance;
-
 NoAggregated BTreeNodeAggregatedWrap<NoAggregated>::_instance;
 template <>
 MinMaxAggregated BTreeNodeAggregatedWrap<MinMaxAggregated>::_instance =
@@ -14,8 +12,6 @@ MinMaxAggregated BTreeNodeAggregatedWrap<MinMaxAggregated>::_instance =
 
 template class BTreeNodeDataWrap<uint32_t, 16>;
 template class BTreeNodeDataWrap<BTreeNoLeafData, 16>;
-template class BTreeKeyData<uint32_t, uint32_t>;
-template class BTreeKeyData<uint32_t, int32_t>;
 template class BTreeNodeT<uint32_t, 16>;
 template class BTreeNodeTT<uint32_t, uint32_t, NoAggregated, 16>;
 template class BTreeNodeTT<uint32_t, BTreeNoLeafData, NoAggregated, 16>;
